feature: poa task split into separate subtasks
