+++ crates/consensus/authority/src/lib.rs
@@ -41,7 +41,9 @@ use std::{collections::HashMap, sync::Arc};
 use voting::{AuthorityVoteCollection, Vote};
 
 use tokio::sync::{RwLock, RwLockReadGuard, RwLockWriteGuard};
-use tracing::{error, info, trace, warn};
+use tracing::{error, trace, warn};
+mod block_builder;
+mod block_fetcher;
 mod builder;
 mod client;
 mod epoch_manager;
